fix bugs and change name `SupportSTDP`
