+++ brainpy/_src/mixin.py
@@ -33,7 +33,7 @@ __all__ = [
   'TreeNode',
   'BindCondData',
   'JointType',
-  'SupportPlasticity',
+  'SupportSTDP',
 ]
 
 global_delay_data = dict()
@@ -562,13 +562,13 @@ class BindCondData(MixIn):
     self._conductance = None
 
 
-class SupportPlasticity(MixIn):
+class SupportSTDP(MixIn):
   """Support synaptic plasticity by modifying the weights.
   """
-  def plasticity(self,
-                 dW: Union[bm.Array, jax.Array],
-                 constraints: Optional[Callable] = None,
-                 ):
+  def update_STDP(self,
+                  dW: Union[bm.Array, jax.Array],
+                  constraints: Optional[Callable] = None,
+                  ):
     raise NotImplementedError
