+++ brainpy/mixin.py
@@ -10,5 +10,5 @@ from brainpy._src.mixin import (
   Container as Container,
   TreeNode as TreeNode,
   JointType as JointType,
-  SupportPlasticity as SupportPlasticity,
+  SupportSTDP as SupportPlasticity,
 )
